+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Descriptions to workflows and steps
 - Added cursor attribute to workflows
 - Basin characteristics computation to "Delineation" workflow
+- Disabled 'next' until user adequately completes the step.
 
 ### Changed  
 
@@ -28,6 +29,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Updated to google analytics 4
 - Burn Severity is turned off by default for Fire Hydrology workflows
 - Some checkboxes are now radio buttons where appropriate
+- 'Query by Fire Perimeters' workflow to 'Query by Fire Perimeter' 
 
 ### Deprecated 
 
@@ -35,7 +37,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Removed 
 
-- 
+- Removed 'Trace Fire Perimeter' step
 
 ### Fixed